src/tests/MGR.CommandLineParser.UnitTests/Converters/CharConverterTests.cs
using System;
using MGR.CommandLineParser.Converters;
using Xunit;

namespace MGR.CommandLineParser.UnitTests.Converters
{
    public class CharConverterTests
    {
        [Fact]
        public void TargetType()
        {
            // Arrange
            IConverter converter = new CharConverter();
            Type expectedType = typeof (Char);

            // Act
            Type actualType = converter.TargetType;

            // Assert
            Assert.Equal(expectedType, actualType);
        }

        [Fact]
        public void Conversion()
        {
            // Arrange
            IConverter converter = new CharConverter();
            string value = "c";
            Char expectedValue = 'c';

            // Act
            object actualValue = converter.Convert(value, converter.TargetType);

            // Assert
            Assert.NotNull(actualValue);
            Assert.IsType<char>(actualValue);
            Assert.Equal(expectedValue, (Char) actualValue);
        }

        [Fact]
        public void BadValueConversion()
        {
            // Arrange
            IConverter converter = new CharConverter();
            string value = "Hello";
            string expectedExceptionMessage = "Unable to parse 'Hello' to Char.";
            string expectedInnerExceptionMessage = "String must be exactly one character long.";

            // Act
            using (new LangageSwitcher("en-us"))
            {
                var actualException = Assert.Throws<CommandLineParserException>(() => converter.Convert(value, converter.TargetType));

                // Assert
                Assert.Equal(expectedExceptionMessage, actualException.Message);
                Assert.NotNull(actualException.InnerException);
                var actualInnerExecption = Assert.IsAssignableFrom<FormatException>(actualException.InnerException);
                Assert.Equal(expectedInnerExceptionMessage, actualInnerExecption.Message);
            }
        }
    }
}